HW-2 : remove NetworkConnectionInterceptor.kt and simplify using MovieApi
