--- class12/dlqkfzlr/study/app/src/main/java/m/woong/architecturestudy/network/MovieApi.kt
@@ -1,35 +1,0 @@
-package m.woong.architecturestudy.network
-
-import m.woong.architecturestudy.network.response.MovieResponse
-import okhttp3.OkHttpClient
-import retrofit2.Call
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import retrofit2.http.GET
-import retrofit2.http.Query
-
-interface MovieApi {
-
-    @GET("v1/search/movie.json")
-    fun movieSearch(
-        @Query("query") keyword: String
-    ): Call<MovieResponse>
-
-    companion object {
-        operator fun invoke(
-            networkConnectionInterceptor: NetworkConnectionInterceptor
-        ): MovieApi {
-
-            val okkHttpclient = OkHttpClient.Builder()
-                .addInterceptor(networkConnectionInterceptor)
-                .build()
-
-            return Retrofit.Builder()
-                .client(okkHttpclient)
-                .baseUrl("https://openapi.naver.com/")
-                .addConverterFactory(GsonConverterFactory.create())
-                .build()
-                .create(MovieApi::class.java)
-        }
-    }
-}
